moves email verification exception to inside the email verification recipe
