+++ src/main/java/io/supertokens/webserver/api/emailverification/VerifyEmailAPI.java
@@ -21,8 +21,8 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import io.supertokens.Main;
 import io.supertokens.emailpassword.User;
-import io.supertokens.emailpassword.exceptions.EmailVerificationInvalidTokenException;
 import io.supertokens.emailverification.EmailVerification;
+import io.supertokens.emailverification.exception.EmailVerificationInvalidTokenException;
 import io.supertokens.pluginInterface.emailpassword.exceptions.UnknownUserIdException;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
