+++ src/test/java/io/supertokens/test/EmailPasswordTest_2.java
@@ -20,9 +20,9 @@ import io.supertokens.ProcessState;
 import io.supertokens.emailpassword.EmailPassword;
 import io.supertokens.emailpassword.User;
 import io.supertokens.emailpassword.UserPaginationContainer;
-import io.supertokens.emailpassword.exceptions.EmailAlreadyVerifiedException;
-import io.supertokens.emailpassword.exceptions.EmailVerificationInvalidTokenException;
 import io.supertokens.emailverification.EmailVerification;
+import io.supertokens.emailverification.exception.EmailAlreadyVerifiedException;
+import io.supertokens.emailverification.exception.EmailVerificationInvalidTokenException;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.pluginInterface.emailpassword.EmailVerificationTokenInfo;
 import io.supertokens.pluginInterface.emailpassword.exceptions.DuplicateEmailVerificationTokenException;
